Move script fulling database in another foldier
